Fix high-priority bug.

+++ src/item.cpp
@@ -545,8 +545,8 @@ Attr_ReadValue Item::readAttr(AttrTypes_t attr, PropStream& propStream)
 		}
 
 		case ATTR_IMBUINGSLOTS: {
-			uint8_t imbuingSlots;
-			if (!propStream.read<uint8_t>(imbuingSlots)) {
+			int32_t imbuingSlots;
+			if (!propStream.read<int32_t>(imbuingSlots)) {
 				return ATTR_READ_ERROR;
 			}
 
@@ -752,7 +752,7 @@ void Item::serializeAttr(PropWriteStream& propWriteStream) const
 
 	if (hasAttribute(ITEM_ATTRIBUTE_IMBUINGSLOTS)) {
 		propWriteStream.write<uint8_t>(ATTR_IMBUINGSLOTS);
-		propWriteStream.write<uint8_t>(getIntAttr(ITEM_ATTRIBUTE_IMBUINGSLOTS));
+		propWriteStream.write<int32_t>(getIntAttr(ITEM_ATTRIBUTE_IMBUINGSLOTS));
 	}
 
 	if (hasAttribute(ITEM_ATTRIBUTE_ARMOR)) {
